refactor: product form interface 위치 이동
